klamath_3c: move resource allocation into pmem_api.
add a pmem_context struct (setup by pmem_api) which keeps track of resources.
struct data_buffer_info is used to keep track of memory buffers allocated for
a given request. it has been moved from work_item into a pmem_api specific
"struct pmem_context". this is the right place for it to be, as resource
allocation is dependent upon the pmem_api type (block or mem).
the pmem_context has also been augmented with a copy of the metadata block.
the latter avoids having to do memory allocation in pmem layer.