parallel restore part 2
On 8 Gbytes NVMe cache and all blocks invalid, restore time single-threaded is 145 seconds. With 64 threads and all blocks invalid, it's 12 seconds, or about 400 Mbytes/second (12x improvement in restore speed). This portion adds parallel restore using a simple workqueue-based scheme. this restore has two advantages over a pure bio-based restore: (1) it's the simplest possible change compared to the current code. (2) it maintains the use of pmem_api -- right now it is not important, but when we begin supporting DIF and 4160-bytes sector sizes, it will be very important to use the pmem_api interfaces, as standard bio restore will not work.